requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
customtkinter
matplotlib
pandas
tkcalendar